Support inline Spark SQL Expressions
